+++ src/Job/Manager.cpp
@@ -1,6 +1,7 @@
 #include "Manager.h"
 #include "Loader.h"
 #include "Printer.h"
+#include "Screenshotter.h"
 #include "ContentProvider.h"
 
 #include "include/wrapper/cef_helpers.h"
@@ -12,6 +13,7 @@ namespace job {
 
 std::size_t Manager::Queue(CefRefPtr<Job> job)
 {
+    std::lock_guard<std::recursive_mutex> lock(m_mutex);
     DLOG(INFO) << "Manager::Queue";
 
     m_jobsQueue.push(job);
@@ -20,6 +22,7 @@ std::size_t Manager::Queue(CefRefPtr<Job> job)
 
 CefRefPtr<CefStreamReader> Manager::GetStreamReader(CefRefPtr<CefBrowser> browser)
 {
+    std::lock_guard<std::recursive_mutex> lock(m_mutex);
     auto it = Find(browser);
     DCHECK(it != m_jobs.end());
 
@@ -32,8 +35,22 @@ CefRefPtr<CefStreamReader> Manager::GetStreamReader(CefRefPtr<CefBrowser> browse
     return it->streamReader;
 }
 
+CefRefPtr<Job> Manager::GetJob(CefRefPtr<CefBrowser> browser)
+{
+    std::lock_guard<std::recursive_mutex> lock(m_mutex);
+    auto it = Find(browser);
+    return it == m_jobs.end() ? nullptr : it->job;
+}
+
+CefString Manager::GetInputMediaType(CefRefPtr<CefBrowser> browser)
+{
+    CefRefPtr<Job> job = GetJob(browser);
+    return job ? job->GetInputMediaType() : CefString("text/html");
+}
+
 void Manager::Assign(CefRefPtr<CefBrowser> browser)
 {
+    std::lock_guard<std::recursive_mutex> lock(m_mutex);
     DLOG(INFO) << "Manager::Assign";
 
     DCHECK(Find(browser) == m_jobs.end());
@@ -46,6 +63,7 @@ void Manager::Assign(CefRefPtr<CefBrowser> browser)
         m_jobs.push_back(BrowserJob({browser, job, nullptr}));
 
         job->SetStatus(Job::Status::LOADING);
+        browser->GetHost()->WasResized();
 
         // Load URL to print
         CefRefPtr<Loader> loader = new Loader(browser->GetMainFrame());
@@ -55,6 +73,7 @@ void Manager::Assign(CefRefPtr<CefBrowser> browser)
 
 void Manager::Process(CefRefPtr<CefBrowser> browser, int httpStatusCode)
 {
+    std::lock_guard<std::recursive_mutex> lock(m_mutex);
     DLOG(INFO) << "Manager::Process";
 
     auto it = Find(browser);
@@ -63,14 +82,18 @@ void Manager::Process(CefRefPtr<CefBrowser> browser, int httpStatusCode)
         if (!httpStatusCode || (200 <= httpStatusCode && 300 > httpStatusCode)) {
             // Generate file name if empty
             if (it->job->GetOutputPath().empty()) {
-                it->job->SetOutputPath(reserveTempFile());
+                it->job->SetOutputPath(reserveTempFile(GetOutputExtension(it->job->GetOutputFormat())));
             }
 
-            it->job->SetStatus(Job::Status::PRINTING);
+            it->job->SetStatus(Job::Status::RENDERING);
 
-            // Print PDF
-            CefRefPtr<Printer> printer = new Printer(this, browser);
-            it->job->accept(printer);
+            if (it->job->GetOutputFormat() == Job::OutputFormat::PDF) {
+                CefRefPtr<Printer> printer = new Printer(this, browser);
+                it->job->accept(printer);
+            } else {
+                CefRefPtr<Screenshotter> screenshotter = new Screenshotter(this, browser, it->job);
+                screenshotter->Start();
+            }
         } else {
             Resolve(it, Job::Status::HTTP_ERROR);
         }
@@ -79,14 +102,16 @@ void Manager::Process(CefRefPtr<CefBrowser> browser, int httpStatusCode)
 
 void Manager::Finish(CefRefPtr<CefBrowser> browser, const CefString& path, bool ok)
 {
+    std::lock_guard<std::recursive_mutex> lock(m_mutex);
     auto it = Find(browser);
     if (it != m_jobs.end()) {
-        Resolve(it, ok ? Job::Status::SUCCESS : Job::Status::PRINT_ERROR);
+        Resolve(it, ok ? Job::Status::SUCCESS : Job::Status::OUTPUT_ERROR);
     }
 }
 
 void Manager::Abort(CefRefPtr<CefBrowser> browser, CefLoadHandler::ErrorCode errorCode)
 {
+    std::lock_guard<std::recursive_mutex> lock(m_mutex);
     auto it = Find(browser);
     if (it != m_jobs.end()) {
         Resolve(it, errorCode == CefLoadHandler::ErrorCode::ERR_ABORTED ? Job::Status::ABORTED : Job::Status::LOAD_ERROR);
@@ -117,6 +142,7 @@ void Manager::Resolve(Manager::Iterator it, const Job::Status& status)
 
 void Manager::StopAll()
 {
+    std::lock_guard<std::recursive_mutex> lock(m_mutex);
     while (!m_jobsQueue.empty()) {
         m_jobsQueue.pop();
     }